% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessMe.R
\name{accuracy}
\alias{accuracy}
\title{Test the robustness of clusters}
\usage{
accuracy(
  accuracy_list = NULL,
  giveassessment = NULL,
  data = NULL,
  cpart = NULL,
  clustsize = NULL,
  crossvali = 50,
  features = NULL,
  ntree = 200,
  loreg = F,
  set.name = NULL
)
}
\arguments{
\item{accuracy_list}{list of accuracy computations/n-fold cross validations, can be utilized in a for loop in order to combine different n-cross-validations for different assessments in one list. Otherwise different n-cross-validations can be combined later to a list which is subjected to the accuracy_plot function.}

\item{data}{count data (un-normalized) from which cells are sampled for cross validation. If \code{NULL}, count data is derived from \code{giveassessment}.}

\item{cpart}{cluster partition. Default = \code{NULL} and partition retrieved from \code{giveassessment}.}

\item{clustsize}{cluster size to be included for cross-validation. Default = \code{NULL} and cluster size is inferred from \code{giveassessment}.}

\item{crossvali}{n number of subsampling to be done for cross-validation(n fold crossvalidation). Default = 50.}

\item{features}{vector of features to be utilized for cross-validation. Default = \code{NULL}.}

\item{ntree}{number of trees to grow for random forest based reclassification. Default = 200.}

\item{loreg}{logical. If \code{T}, perform n-fold crossvalidation based on multionomial logistic regression. Default = \code{F}.}

\item{set.name}{set name of list object for n-fold crossvalidation within accuracy list. Default = \code{NULL}.}

\item{giveassesment}{assessment object}
}
\value{
accuracy list with objects/lists for different accuracy computations e.g. for different cluster partitions. Every object representing a list with n-fold crossvalidation objects with fractions of reclassified/re-labelled cells matching the original cluster label with re-classification based on different gene lists derived from \code{giveassessment}.
}
\description{
This function serves to explore the robustness of clusters of the cluster partition based on enriched genes, unique enriched genes, outlier genes, feature genes as well as, if assessed, shared enriched but differentially expressed genes.
}
\examples{
accuracy_list <- accuracy(giveassessment = assess_seuratRC$Sres.1, data = entero@assays$RNA@counts)
accuracy_list <- accuracy(accuracy_list = accuracy_list,giveassessment = assess_seuratRC$Sres.6, data = entero@assays$RNA@counts)
}
