% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessMe.R
\name{accuracy_plot}
\alias{accuracy_plot}
\title{Plot and compare the robustness of different clusters partitions}
\usage{
accuracy_plot(accuracy_list)
}
\arguments{
\item{accuracy_list}{list of accuracy computations/n-fold cross validations for e.g. different cluster partitions.}
}
\description{
This function plots the accuracy of re-classifications of the different cluster partitions tested based on enriched genes, unique enriched genes, outlier genes, feature genes as well as if assessed shared enriched but differentially expressed genes.
}
\examples{
accuracy_plot(accuracy_list)
}
