% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessMe.R
\name{generate_input_accuracy_hpc}
\alias{generate_input_accuracy_hpc}
\title{Generate input for accuracy function on HPC systems}
\usage{
generate_input_accuracy_hpc(
  assessment,
  object = NULL,
  crossvali = NULL,
  assessment_no = NULL,
  rawdata = NULL,
  tosample = NULL
)
}
\arguments{
\item{assessment}{assessment object from \code{cluster_assessment} function.}

\item{object}{RaceID or Seurat object from which count data is derived. Default = \code{NULL}.}

\item{crossvali}{n number of subsampling to be done for cross-validation(n fold crossvalidation). Default = \code{NULL}.}

\item{assessment_no}{index of assessment object within a list of assessments. Default = \code{NULL}.}

\item{rawdata}{count data. Default = \code{NULL}.}

\item{tosample}{percentage of cells to sample for the training set, per default 70\% are sampled. Default = \code{NULL}.}
}
\description{
helper function to generate input for accuracy function executed on a high performing computing cluster.
}
\examples{
tes_generate <- generate_input_accuracy(assessment=assessment_input, object=giveobject, crossvali=csvali, assessment_no = assess_index, rawdata = raw, tosample=sample2 )
}
