% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessMe.R
\name{one_plot_output_fix_new}
\alias{one_plot_output_fix_new}
\title{Plot function to determine optimal pre-processing method.}
\usage{
one_plot_output_fix_new(
  assesment_list2,
  cex = 1,
  lcol = c("red"),
  map = T,
  leg = T
)
}
\arguments{
\item{cex}{= numeric, graphical parameter indicating the amount by which the line connecting the data points should be scaled. Default = 1}

\item{lcol}{= vector of colors used for highlighting slots of list of assessments, for each list of lists of assessment one color: e.g. list of resolution optimizations for e.g. normalization A, and another color for list of resolution optimization for e.g. normalization B.}

\item{map}{= logical. If \code{T}, then legend is shown. Default = \code{T}.}

\item{leg}{= logical. If \code{T}, then the legend is shown. Default = \code{T}.}

\item{assessment_list2}{list of assessment lists exhibiting different assessments, for example: different normalization methods and for each increasing resolution parameters.}

\item{f1_thr}{numeric, threshold used to calculate how many clusters have at least 1 gene with F1-score above this threshold for different cluster partitions assessed. Default = 0.5.}

\item{max_leng}{numeric, calculation of number of clusters with at least \code{max_leng} genes with minimal F1-score of \code{f1_thr} for the different cluster partitions assessed. Default = 3.}
}
\description{
After cluster assessment, this function serves to identify optimal pre-processing method, independent of the number of clusters, plotting the following criteria: average F1-Score, average Entropy, average number of outlier genes across cluster and average number of enriched features across clusters.
}
\examples{
one_plot_output_fix_new(assessment_list, lcol = c("red", "blue", "green"))
}
