% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessMe.R
\name{opti_preprocess_plot}
\alias{opti_preprocess_plot}
\title{Plot function to determine optimal pre-processing method.}
\usage{
opti_preprocess_plot(
  assesment_list2,
  cex = 1,
  lcol = c("red"),
  map = T,
  leg = T
)
}
\arguments{
\item{cex}{= numeric, graphical parameter indicating the amount by which the line connecting the data points should be scaled. Default = 1}

\item{lcol}{= vector of colors used for highlighting slots of list of assessments, for each list of lists of assessment one color: e.g. list of resolution optimizations for e.g. normalization A, and another color for list of resolution optimization for e.g. normalization B.}

\item{map}{= logical. If \code{T}, then legend is shown. Default = \code{T}.}

\item{leg}{= logical. If \code{T}, then the legend is shown. Default = \code{T}.}

\item{assessment_list2}{list of assessment lists exhibiting different assessments, for example: different normalization methods and for each increasing resolution parameters.}

\item{f1_thr}{numeric, threshold used to calculate how many clusters have at least 1 gene with F1-score above this threshold for different cluster partitions assessed. Default = 0.5.}

\item{max_leng}{numeric, calculation of number of clusters with at least \code{max_leng} genes with minimal F1-score of \code{f1_thr} for the different cluster partitions assessed. Default = 3.}
}
\value{
plot with 6 graphs, plotting information about cluster partition against number of clusters and number of assessed genes, as well as plotting number of clusters against average F1-score, average Entropy, average number of enriched features assessed and average No. of outlier genes across clusters.
\item{output_list}{with with data.frame for every list of assessments, with with different resolutions/cluster partitions as rows and the following columns: “No.cluster” = number of assessed clusters, “mean_F1” = mean F1_Score across genes, “mean_Entropy” = mean Entropy across genes, “mean_No.cell_outlg1” = mean number of cells with 1 outlier gene expression across clusters, "mean_enriched_features" = mean numer of enriched feautres across clusters of assessed features and "assessed_features" = Number of assessed features.}
}
\description{
After cluster assessment, this function serves to identify optimal pre-processing method, independent of the number of clusters, plotting the following criteria: average F1-Score, average Entropy, average number of outlier genes across cluster and average number of enriched features across clusters.
}
\examples{
opti_preprocess_plot(assessment_list, lcol = c("red", "blue", "green"))
}
