% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessMe.R
\name{plotfit_new}
\alias{plotfit_new}
\title{differences in f1-score or entropy of individual genes}
\usage{
plotfit_new(
  assessment1,
  assessment2,
  maplot = F,
  toplot = "F1",
  signifi = F,
  out_doub_dif = NULL,
  xlabs = NULL,
  ylabs = NULL,
  logmean = NULL
)
}
\arguments{
\item{assessment1}{assessment of a cluster partition for which either f1_score or Entropy computation has been performed based on a derived per gene cutoff.}

\item{assessment2}{assessment of a second partition derived for the same data.}

\item{maplot}{logical. If \code{T}, plot the F1 or Entropy per gene difference of assessment 1 and assessment 2 against the mean expression of the gene.}

\item{toplot}{plot either “F1” or “Entropy”. Differences of F1 or Entropy calculations of genes between the assessed cluster partitions. Default = "F1".}

\item{signifi}{logical. if \code{T}, calculate standard deviation of absolute differences between F1/Entropy gene-pairs of two assessments and multiplies this standard deviation by 1.64485361, representing the z-score with significance level based on the normal distribution function. The expected difference of a pair equals 0. Genes with 1.64485361 standard deviations(sds) difference between the two assessments are considered significant and are highlighted. Default = \code{F} and all genes with differences of one standard deviation are highlighted.}

\item{out_doub_dif}{vector of gene names of user specified gene sets to highlight. Default = \code{NULL} and genes with differences larger than 1 sd or 1.64485361 sds are highlighted.}

\item{xlabs}{character string for x-axis label indicating assessment 1. Default = \code{NULL} and label is derived from object name of assessment 1.}

\item{ylabs}{character string for y-axis label indicating assessment 2. Default = \code{NULL} and label is derived from object name of assessment 2.}

\item{logmean}{either “log2”, “log10” or NULL. If “log2” or “log10” performs log transformation of the mean expression per gene. Default = \code{NULL}.}
}
\description{
This function serves to explore differences in f1-score or entropy of individual genes between different assessed cluster partitions. Genes with largest differences between the two assessments are highlighted.
}
\examples{
plotfit_new(assessmentRC$Sres.1, assessmentRC$Sres.5, maplot = T, logmean = "log2", signifi = T)
}
